refactor to use existing version of session storage service
