+++ frontend/src/components/LoginModal.tsx
@@ -1,6 +1,6 @@
 "use-client";
 
-import SessionStorage from "src/services/auth/sessionStorage";
+import SessionStorage from "src/services/sessionStorage/sessionStorage";
 
 import { RefObject } from "react";
 import {
