+++ frontend/src/components/ReturnToGrantsNotification.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { useSessionStorage } from "src/services/sessionStorage/useSessionStorage";
+import SessionStorage from "src/services/sessionStorage/sessionStorage";
 
 import { useTranslations } from "next-intl";
 import { useSearchParams } from "next/navigation";
 
 export function ReturnToGrantsNotification() {
   const t = useTranslations("returnToGrants");
-  const { getSessionStorageItem, setSessionStorageItem } = useSessionStorage();
   const searchParams = useSearchParams();
   const GrantsLink = (
     <div className="display-flex flex-1 tablet:text-right tablet:margin-bottom-0 margin-bottom-3 tablet:padding-top-3 tablet:flex-justify-end">
@@ -15,10 +14,10 @@ export function ReturnToGrantsNotification() {
     </div>
   );
   if (searchParams.get("utm_source") === "Grants.gov") {
-    setSessionStorageItem("showLegacySearchReturnNotification", "true");
+    SessionStorage.setItem("showLegacySearchReturnNotification", "true");
     return GrantsLink;
   }
-  if (getSessionStorageItem("showLegacySearchReturnNotification") === "true") {
+  if (SessionStorage.getItem("showLegacySearchReturnNotification") === "true") {
     return GrantsLink;
   }
