+++ frontend/tests/pages/subscribe/page.test.tsx
@@ -4,6 +4,9 @@ import { identity } from "lodash";
 import Subscribe from "src/app/[locale]/subscribe/page";
 import { localeParams, useTranslationsMock } from "src/utils/testing/intlMocks";
 
+const mockSetItem = jest.fn();
+const mockGetItem = jest.fn();
+
 jest.mock("react-dom", () => {
   const originalModule =
     jest.requireActual<typeof import("react-dom")>("react-dom");
@@ -34,12 +37,16 @@ jest.mock("next/navigation", () => ({
   useSearchParams: () => new URLSearchParams(),
 }));
 
-jest.mock("src/services/sessionStorage/useSessionStorage", () => ({
-  useSessionStorage: () => ({
-    getSessionStorageItem: jest.fn(),
-    setSessionStorageItem: jest.fn(),
-  }),
-}));
+jest.mock("src/services/sessionStorage/sessionStorage", () => {
+  return {
+    __esModule: true,
+    default: {
+      setItem: (key: string, value: string) =>
+        mockSetItem(key, value) as unknown,
+      getItem: (key: string) => mockGetItem(key) as unknown,
+    },
+  };
+});
 
 describe("Subscribe", () => {
   it("renders intro text", () => {
